-간단한 버그 픽스완료

+++ backend/MyDiary/src/main/java/com/example/MyDiary/Entity/AI_RepliesEntity.java
@@ -15,7 +15,7 @@ import lombok.*;
 public class AI_RepliesEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer replyId;
+    private Long replyId;
 
     @OneToOne
     @MapsId
